Refactor get/curSharedUsers to redux
