+++ client/src/components/navbar.jsx
@@ -13,44 +13,44 @@ class NavBar extends React.Component {
   };
 
 // Probably, the logic for getting shared users should be an effect of textEditor mounting
-  componentWillMount() {
-    // FOR GETTING DOCID & USERID FROM STATE // ERROR: NOT GETTING SET IN TIME TO USE THEM HERE, THEY ARE NULL
-    // var docId = this.props.curDoc;
-    // var userId = this.props.userId;
+  // componentWillMount() {
+  //   // FOR GETTING DOCID & USERID FROM STATE // ERROR: NOT GETTING SET IN TIME TO USE THEM HERE, THEY ARE NULL
+  //   // var docId = this.props.curDoc;
+  //   // var userId = this.props.userId;
 
-    // GET USERID + DOCID => SHARED USERS ON THE DOC // BANDAID FOR NOT GETTING VALUES FROM STATE
-    var urldocId = window.location.search.split('').splice(11).join('');
-    var username = window.localStorage.user.slice(1, window.localStorage.user.length - 1);
-    var sharelinkId = urldocId.length === 0 ? 'hr46' : urldocId; // default to public doc if there is no doc id in url
-    var docId = null;
+  //   // GET USERID + DOCID => SHARED USERS ON THE DOC // BANDAID FOR NOT GETTING VALUES FROM STATE
+  //   var urldocId = window.location.search.split('').splice(11).join('');
+  //   var username = window.localStorage.user.slice(1, window.localStorage.user.length - 1);
+  //   var sharelinkId = urldocId.length === 0 ? 'hr46' : urldocId; // default to public doc if there is no doc id in url
+  //   var docId = null;
 
-    $.ajax({
-      method: 'GET',
-      url: '/document/id',
-      dataType: 'json',
-      data: {
-        sharelinkId: sharelinkId
-      },
-      success: (data) => {
-        console.log('NAVBAR found docID from sharelink:', data);
-        var docId = data.id;
-        axios.get('users/id/?username=' + username)
-          .then(function(res) {
-            console.log('NAVBAR success getting userId:', res.data);
-            var userId = res.data;
-            this.props.setUserId(userId);
-            this.props.getSharedUsers(docId, userId);
-          }.bind(this))
-          .catch(function(err) {
-            console.log('NAVBAR Error retrieving user.')
-          });
-      },
-      error: (err) => {
-        console.log('TVP getDocId error:', err);
-      }
-    })
+  //   $.ajax({
+  //     method: 'GET',
+  //     url: '/document/id',
+  //     dataType: 'json',
+  //     data: {
+  //       sharelinkId: sharelinkId
+  //     },
+  //     success: (data) => {
+  //       console.log('NAVBAR found docID from sharelink:', data);
+  //       var docId = data.id;
+  //       axios.get('users/id/?username=' + username)
+  //         .then(function(res) {
+  //           console.log('NAVBAR success getting userId:', res.data);
+  //           var userId = res.data;
+  //           this.props.setUserId(userId);
+  //           this.props.getSharedUsers(docId, userId);
+  //         }.bind(this))
+  //         .catch(function(err) {
+  //           console.log('NAVBAR Error retrieving user.')
+  //         });
+  //     },
+  //     error: (err) => {
+  //       console.log('TVP getDocId error:', err);
+  //     }
+  //   })
 
-  }
+  // }
 
   render() {
     return (
@@ -103,7 +103,8 @@ function mapStateToProps(state) {
     docId: state.tvPage.curDoc,
     // userId: state.documentlist.curUser,
     userId: state.navbar.userId,
-    sharelink: state.editor.sharelinkId
+    sharelink: state.editor.sharelinkId,
+    curSharedUsers: state.tvPage.curSharedUsers
   }
 }
