+++ client/src/reducers/tvPageReducer.jsx
@@ -1,6 +1,7 @@
 const initialState = {
   curUser: null,
-  curDoc: null
+  curDoc: null,
+  curSharedUsers: []
 }
 
 export default function (state=initialState, action) {
@@ -9,6 +10,10 @@ export default function (state=initialState, action) {
     case 'SET_DOC_ID':
       return Object.assign({}, state, {
         curDoc: action.docId
+      })
+    case 'SET_CUR_SHARED_USERS':
+      return Object.assign({}, state, {
+        curSharedUsers: action.users
       })
     default:
       return state
